% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vue.R
\name{Vue}
\alias{Vue}
\title{Create a Vue App}
\usage{
Vue(html, data = list(), dependencies = list(), ...,
  elementId = NULL)
}
\description{
Create a Vue App
}
\examples{
tags$div(
  tags$label('Enter your name'),
  tags$input(type = "text", "v-model" = "name"),
  tags$p("Hello {{name}}")
  ) \%>\%
  Vue(data = list(name = ""))
}
